Add files via upload
Make sure all three files are in the same directory.

When you open the index.html file in a web browser, you'll see a canvas element displaying the game. The player can control the character using the arrow keys to avoid the red obstacle. If the character collides with the obstacle, a "Game Over!" alert will appear, and the game will reset.